requirements.txt
transformers>=4.39.0
streamlit>=1.34.0
plotly>=5.21.0
wordcloud>=1.9.3
matplotlib>=3.9.0
nltk>=3.9
python-dotenv>=1.0.1
scikit-learn>=1.4.1
datasets>=2.18.0
pandas>=2.2.1
PyYAML>=6.0.1
